cleanup cardstory
