+++ src/gui/src/components/assess/CardStory.vue
@@ -92,6 +92,7 @@
               </v-btn>
 
               <v-btn
+                v-if="!detailView"
                 small
                 class="item-action-btn open-close-btn"
                 :class="openSummary ? 'opened' : 'closed'"
@@ -208,13 +209,9 @@
                     </div>
                   </v-col>
                 </v-row>
-                <v-row>
-                  <LineChart
-                    v-if="openSummary && !published_date_outdated"
-                    :chart-options="chartOptions"
-                    :chart-data="chart_data"
-                    :width="400"
-                    :height="200"
+                <v-row v-if="openSummary && !published_date_outdated">
+                  <week-chart
+                    :story="story"
                   />
                 </v-row>
                 <metainfo
@@ -257,35 +254,14 @@ import PopupDeleteItem from '@/components/popups/PopupDeleteItem'
 import PopupShareItems from '@/components/popups/PopupShareItems'
 import metainfo from '@/components/assess/card/metainfo'
 import CardNewsItem from '@/components/assess/CardNewsItem'
+import WeekChart from '@/components/assess/card/WeekChart'
 
 import { mapGetters } from 'vuex'
-import { Line as LineChart } from 'vue-chartjs/legacy'
-import {
-  Chart as ChartJS,
-  Title,
-  Tooltip,
-  Legend,
-  LineElement,
-  LinearScale,
-  CategoryScale,
-  PointElement
-} from 'chart.js'
-
 import {
   deleteNewsItemAggregate,
   importantNewsItemAggregate,
   readNewsItemAggregate
 } from '@/api/assess'
-
-ChartJS.register(
-  Title,
-  Tooltip,
-  Legend,
-  LineElement,
-  LinearScale,
-  CategoryScale,
-  PointElement
-)
 
 export default {
   name: 'CardStory',
@@ -294,7 +270,7 @@ export default {
     PopupDeleteItem,
     PopupShareItems,
     metainfo,
-    LineChart
+    WeekChart
   },
   emits: ['selectItem', 'deleteItem'],
   props: {
@@ -302,20 +278,18 @@ export default {
       type: Object,
       required: true
     },
-    selected: Boolean,
+    selected: { type: Boolean, default: false },
     detailView: { type: Boolean, default: false }
   },
-  data: () => ({
-    viewDetails: false,
-    openSummary: false,
-    sharingDialog: false,
-    deleteDialog: false,
-    showDialog: false,
-    chartOptions: {
-      responsive: true,
-      maintainAspectRatio: false
+  data: function () {
+    return {
+      viewDetails: false,
+      openSummary: this.detailView,
+      sharingDialog: false,
+      deleteDialog: false,
+      showDialog: false
     }
-  }),
+  },
   computed: {
     item_important() {
       return 'important' in this.story ? this.story.important : false
@@ -344,48 +318,6 @@ export default {
     story_in_report() {
       return this.story.in_reports_count > 0
     },
-
-    last_seven_days() {
-      return Array.from(Array(7).keys(), (i) => {
-        const date = new Date()
-        date.setDate(date.getDate() - i)
-        return date.toLocaleDateString(undefined, { day: '2-digit', month: '2-digit' })
-      })
-    },
-
-    news_items_per_day() {
-      const days = this.last_seven_days
-      const items_per_day = this.story.news_items.reduce((acc, item) => {
-        const day = new Date(item.news_item_data.published).toLocaleDateString(undefined, { day: '2-digit', month: '2-digit' })
-        if (day in acc) {
-          acc[day] += 1
-        } else {
-          acc[day] = 1
-        }
-        return acc
-      }, {})
-
-      return days.map(day => {
-        if (day in items_per_day) {
-          return items_per_day[day]
-        } else {
-          return 0
-        }
-      })
-    },
-
-    chart_data() {
-      return {
-        labels: this.last_seven_days,
-        datasets: [
-          {
-            label: 'Anzahl der Artikel',
-            data: this.news_items_per_day
-          }
-        ]
-      }
-    },
-
     news_item_length() {
       return this.story.news_items.length
     },
@@ -442,7 +374,9 @@ export default {
     },
 
     getDescription() {
-      return this.openSummary ? this.story.description : (this.story.summary || this.story.description)
+      return this.openSummary
+        ? this.story.description
+        : this.story.summary || this.story.description
     },
 
     getTags() {
