+++ src/gui/src/i18n/de/messages.js
@@ -403,262 +403,6 @@ export const messages_de = {
     select_datetime: 'Wählen Sie Datum/Uhrzeit aus',
     done: 'Erledigt'
   },
-  cvss_calculator: {
-    title: 'CVSS-Rechner 3.1',
-    base_score: 'Grundpunktzahl',
-    attack_vector: 'Angriffsvektor (AV)',
-    attack_complexity: 'Angriffskomplexität (AC)',
-    privileges_required: 'Erforderliche Berechtigungen (PR)',
-    user_interaction: 'Benutzerinteraktion (UI)',
-    scope: 'Umfang (S)',
-    confidentiality: 'Vertraulichkeit (C)',
-    integrity: 'Integrität (I)',
-    availability: 'Verfügbarkeit (A)',
-    temporal_score: 'Zeitlicher Score',
-    exploitability_code_maturity: 'Reifegrad des Ausnutzbarkeitscodes (E)',
-    remediation_level: 'Sanierungsgrad (RL)',
-    report_confidence: 'Vertrauen melden (RC)',
-    environmental_score: 'Umwelt-Score',
-    confidentiality_requirement: 'Vertraulichkeitsanforderung (CR)',
-    integrity_requirement: 'Integritätsanforderung (IR)',
-    availability_requirement: 'Verfügbarkeitsanforderung (AR)',
-    modified_attack_vector: 'Modifizierter Angriffsvektor (MAV)',
-    modified_attack_complexity: 'Modifizierte Angriffskomplexität (MAC)',
-    modified_privileges_required: 'Geänderte Berechtigungen erforderlich (MPR)',
-    modified_user_interaction: 'Modifizierte Benutzerinteraktion (MUI)',
-    modified_scope: 'Modifizierter Geltungsbereich (MS)',
-    modified_confidentiality: 'Modifizierte Vertraulichkeit (MC)',
-    modified_integrity: 'Modifizierte Integrität (MI)',
-    modified_availability: 'Geänderte Verfügbarkeit (MA)',
-    network: 'Netzwerk',
-    adjacent: 'Benachbart',
-    adjacent_network: 'Angrenzendes Netzwerk',
-    local: 'Lokal',
-    physical: 'Physisch',
-    required: 'Erforderlich',
-    unchanged: 'Unverändert',
-    changed: 'Geändert',
-    not_defined: 'Nicht definiert',
-    none: 'Keiner',
-    low: 'Niedrig',
-    medium: 'Mittel',
-    high: 'Hoch',
-    critical: 'Kritisch',
-    unproven: 'Unbewiesen',
-    proof_of_concept: 'Konzeptioneller Beweiß',
-    functional: 'Funktionell',
-    official_fix: 'Offizielle Lösung',
-    temporary_fix: 'Temporäre Lösung',
-    workaround: 'Problemumgehung',
-    unavailable: 'Nicht verfügbar',
-    unknown: 'Unbekannt',
-    reasonable: 'Vernünftig',
-    confirmed: 'Bestätigt',
-    validator: 'Ungültige oder unvollständige Vektorzeichenfolge'
-  },
-  cvss_calculator_tooltip: {
-    baseMetricGroup_Legend:
-      'Die Basismetrikgruppe stellt die intrinsischen Merkmale einer Schwachstelle dar, die im Laufe der Zeit und über Benutzerumgebungen hinweg konstant sind. ',
-    AV_Heading:
-      'Diese Metrik spiegelt den Kontext wider, in dem die Ausnutzung von Schwachstellen möglich ist. ',
-    AV_N_Label:
-      'Die verwundbare Komponente ist an den Netzwerkstack gebunden und die Menge möglicher Angreifer erstreckt sich über die anderen aufgeführten Optionen hinaus bis hin zum gesamten Internet. ',
-    AV_A_Label:
-      'Die verwundbare Komponente ist an den Netzwerkstapel gebunden, aber der Angriff ist auf Protokollebene auf eine logisch benachbarte Topologie beschränkt. ',
-    AV_L_Label:
-      'Die verwundbare Komponente ist nicht an den Netzwerkstapel gebunden und der Weg des Angreifers führt über Lese-/Schreib-/Ausführungsfunktionen. ',
-    AV_P_Label:
-      'Der Angriff erfordert, dass der Angreifer die verwundbare Komponente physisch berührt oder manipuliert. ',
-    AC_Heading:
-      'Diese Metrik beschreibt die Bedingungen, die außerhalb der Kontrolle des Angreifers liegen müssen, um die Schwachstelle auszunutzen. ',
-    AC_L_Label:
-      'Besondere Zugangsbedingungen oder mildernde Umstände bestehen nicht. ',
-    AC_H_Label:
-      'Ein erfolgreicher Angriff hängt von Bedingungen ab, die außerhalb der Kontrolle des Angreifers liegen. ',
-    PR_Heading:
-      'Diese Metrik beschreibt die Berechtigungsstufe, die ein Angreifer besitzen muss, bevor er die Schwachstelle erfolgreich ausnutzen kann.',
-    PR_N_Label:
-      'Der Angreifer ist vor dem Angriff nicht autorisiert und benötigt daher keinen Zugriff auf Einstellungen oder Dateien, um einen Angriff durchzuführen.',
-    PR_L_Label:
-      'Der Angreifer ist mit (d. h. erfordert) Privilegien autorisiert, die grundlegende Benutzerfunktionen bereitstellen, die normalerweise nur Einstellungen und Dateien betreffen, die einem Benutzer gehören. ',
-    PR_H_Label:
-      'Der Angreifer ist mit (d. h. erforderlichen) Privilegien autorisiert, die eine erhebliche (z. B. administrative) Kontrolle über die anfällige Komponente bieten, die sich auf komponentenweite Einstellungen und Dateien auswirken könnte.',
-    UI_Heading:
-      'Diese Metrik erfasst die Anforderung an einen anderen Benutzer als den Angreifer, sich an der erfolgreichen Kompromittierung der anfälligen Komponente zu beteiligen. ',
-    UI_N_Label:
-      'Das anfällige System kann ohne Interaktion eines Benutzers ausgenutzt werden.',
-    UI_R_Label:
-      'Die erfolgreiche Ausnutzung dieser Schwachstelle erfordert, dass ein Benutzer einige Maßnahmen ergreift, bevor die Schwachstelle ausgenutzt werden kann.',
-    S_Heading:
-      'Wirkt sich ein erfolgreicher Angriff auf eine andere Komponente als die anfällige Komponente aus? ',
-    S_U_Label:
-      'Eine ausgenutzte Schwachstelle kann nur Ressourcen betreffen, die von derselben Sicherheitsbehörde verwaltet werden. ',
-    S_C_Label:
-      'Eine ausgenutzte Schwachstelle kann Ressourcen beeinträchtigen, die über den Sicherheitsbereich hinausgehen, der von der Sicherheitsautorität der anfälligen Komponente verwaltet wird. ',
-    C_Heading:
-      'Diese Metrik misst die Auswirkungen auf die Vertraulichkeit der von einer Softwarekomponente verwalteten Informationsressourcen aufgrund einer erfolgreich ausgenutzten Schwachstelle. ',
-    C_N_Label:
-      'Es gibt keinen Verlust der Vertraulichkeit innerhalb der betroffenen Komponente.',
-    C_L_Label: 'Es gibt einen gewissen Verlust an Vertraulichkeit. ',
-    C_H_Label:
-      'Es kommt zu einem vollständigen Verlust der Vertraulichkeit, was dazu führt, dass alle Ressourcen innerhalb der betroffenen Komponente an den Angreifer weitergegeben werden. ',
-    I_Heading:
-      'Diese Metrik misst die Auswirkung auf die Integrität einer erfolgreich ausgenutzten Schwachstelle. ',
-    I_N_Label:
-      'Es gibt keinen Integritätsverlust innerhalb der betroffenen Komponente.',
-    I_L_Label:
-      'Die Änderung von Daten ist möglich, aber der Angreifer hat keine Kontrolle über die Folgen einer Änderung oder der Umfang der Änderung ist begrenzt. ',
-    I_H_Label:
-      'Es kommt zu einem totalen Integritätsverlust oder einem kompletten Schutzverlust. ',
-    A_Heading:
-      'Diese Metrik misst die Auswirkung auf die Verfügbarkeit der betroffenen Komponente, die sich aus einer erfolgreich ausgenutzten Schwachstelle ergibt. ',
-    A_N_Label:
-      'Es gibt keine Auswirkungen auf die Verfügbarkeit innerhalb der betroffenen Komponente.',
-    A_L_Label:
-      'Die Leistung wird reduziert oder es gibt Unterbrechungen in der Ressourcenverfügbarkeit. ',
-    A_H_Label:
-      'Es kommt zu einem vollständigen Verlust der Verfügbarkeit, was dazu führt, dass der Angreifer den Zugriff auf Ressourcen in der betroffenen Komponente vollständig verweigern kann; ',
-    temporalMetricGroup_Legend:
-      'Die zeitlichen Metriken messen den aktuellen Stand von Exploit-Techniken oder die Verfügbarkeit von Code, das Vorhandensein von Patches oder Problemumgehungen oder das Vertrauen, das man in die Beschreibung einer Schwachstelle hat.',
-    E_Heading:
-      'Diese Metrik misst die Wahrscheinlichkeit, dass die Schwachstelle angegriffen wird, und basiert in der Regel auf dem aktuellen Stand der Exploit-Techniken, der Verfügbarkeit von Exploit-Code oder der aktiven „in-the-wild“-Ausnutzung.',
-    E_X_Label:
-      'Die Zuweisung dieses Werts weist darauf hin, dass nicht genügend Informationen vorhanden sind, um einen der anderen Werte auszuwählen, und hat keinen Einfluss auf die zeitliche Gesamtbewertung, d. h. es hat die gleiche Auswirkung auf die Bewertung wie die Zuweisung von „Hoch“.',
-    E_U_Label:
-      'Es ist kein Exploit-Code verfügbar oder ein Exploit ist theoretisch.',
-    E_P_Label:
-      'Proof-of-Concept-Exploit-Code ist verfügbar oder eine Angriffsdemonstration ist für die meisten Systeme nicht praktikabel. ',
-    E_F_Label: 'Funktionaler Exploit-Code ist verfügbar. ',
-    E_H_Label:
-      'Funktionaler autonomer Code ist vorhanden oder es ist kein Exploit erforderlich (manueller Auslöser) und Details sind allgemein verfügbar. ',
-    RL_Heading:
-      'Der Remediation Level einer Schwachstelle ist ein wichtiger Faktor für die Priorisierung. ',
-    RL_X_Label:
-      'Die Zuweisung dieses Werts zeigt an, dass nicht genügend Informationen vorhanden sind, um einen der anderen Werte auszuwählen, und hat keine Auswirkung auf die zeitliche Gesamtbewertung, d. h. es hat die gleiche Auswirkung auf die Bewertung wie die Zuweisung von „Nicht verfügbar“.',
-    RL_O_Label: 'Eine vollständige Anbieterlösung ist verfügbar. ',
-    RL_T_Label: 'Es gibt eine offizielle, aber vorübergehende Lösung. ',
-    RL_W_Label:
-      'Es gibt eine inoffizielle Lösung, die nicht vom Hersteller stammt. ',
-    RL_U_Label:
-      'Es ist entweder keine Lösung verfügbar oder die Anwendung ist nicht möglich.',
-    RC_Heading:
-      'Diese Metrik misst den Grad des Vertrauens in die Existenz der Schwachstelle und die Glaubwürdigkeit der bekannten technischen Details. ',
-    RC_X_Label:
-      'Die Zuweisung dieses Werts weist darauf hin, dass nicht genügend Informationen vorhanden sind, um einen der anderen Werte auszuwählen, und hat keinen Einfluss auf die zeitliche Gesamtbewertung, d. h. es hat die gleiche Auswirkung auf die Bewertung wie die Zuweisung von „Bestätigt“.',
-    RC_U_Label:
-      'Es gibt Berichte über Auswirkungen, die darauf hindeuten, dass eine Schwachstelle vorhanden ist. ',
-    RC_R_Label:
-      'Signifikante Details werden veröffentlicht, aber die Forscher haben entweder kein volles Vertrauen in die Grundursache oder haben keinen Zugriff auf den Quellcode, um alle Wechselwirkungen, die zu dem Ergebnis führen können, vollständig zu bestätigen. ',
-    RC_C_Label:
-      'Es liegen detaillierte Berichte vor oder eine funktionale Reproduktion ist möglich (funktionale Exploits können dies bereitstellen). ',
-    environmentalMetricGroup_Legend:
-      'Diese Metriken ermöglichen es dem Analysten, den CVSS-Score abhängig von der Bedeutung des betroffenen IT-Assets für das Unternehmen eines Benutzers anzupassen, gemessen in Bezug auf vorhandene ergänzende/alternative Sicherheitskontrollen, Vertraulichkeit, Integrität und Verfügbarkeit. ',
-    CR_Heading:
-      'Diese Metriken ermöglichen es dem Analysten, den CVSS-Score abhängig von der Bedeutung der Vertraulichkeit des betroffenen IT-Assets für die Organisation eines Benutzers im Verhältnis zu anderen Auswirkungen anzupassen. ',
-    CR_X_Label:
-      'Die Zuweisung dieses Werts weist darauf hin, dass nicht genügend Informationen vorhanden sind, um einen der anderen Werte auszuwählen, und hat keinen Einfluss auf die Gesamtbewertung für die Umwelt, d. h. es hat die gleiche Auswirkung auf die Bewertung wie die Zuweisung von Mittel.',
-    CR_L_Label:
-      'Der Verlust der Vertraulichkeit hat wahrscheinlich nur begrenzte nachteilige Auswirkungen auf die Organisation oder Personen, die mit der Organisation verbunden sind (z. B. Mitarbeiter, Kunden).',
-    CR_M_Label:
-      'Das Zuweisen dieses Werts zur Metrik hat keinen Einfluss auf die Punktzahl.',
-    CR_H_Label:
-      'Der Verlust der Vertraulichkeit hat wahrscheinlich katastrophale nachteilige Auswirkungen auf die Organisation oder Personen, die mit der Organisation verbunden sind (z. B. Mitarbeiter, Kunden).',
-    IR_Heading:
-      'Diese Metriken ermöglichen es dem Analysten, den CVSS-Score abhängig von der Bedeutung der Integrität des betroffenen IT-Assets für die Organisation eines Benutzers im Verhältnis zu anderen Auswirkungen anzupassen. ',
-    IR_X_Label:
-      'Die Zuweisung dieses Werts weist darauf hin, dass nicht genügend Informationen vorhanden sind, um einen der anderen Werte auszuwählen, und hat keinen Einfluss auf die Gesamtbewertung für die Umwelt, d. h. es hat die gleiche Auswirkung auf die Bewertung wie die Zuweisung von Mittel.',
-    IR_L_Label:
-      'Der Verlust der Integrität hat wahrscheinlich nur begrenzte nachteilige Auswirkungen auf die Organisation oder Personen, die mit der Organisation verbunden sind (z. B. Mitarbeiter, Kunden).',
-    IR_M_Label:
-      'Das Zuweisen dieses Werts zur Metrik hat keinen Einfluss auf die Punktzahl.',
-    IR_H_Label:
-      'Der Verlust der Integrität hat wahrscheinlich katastrophale nachteilige Auswirkungen auf die Organisation oder Personen, die mit der Organisation verbunden sind (z. B. Mitarbeiter, Kunden).',
-    AR_Heading:
-      'Diese Metriken ermöglichen es dem Analysten, den CVSS-Score abhängig von der Bedeutung der Verfügbarkeit des betroffenen IT-Assets für die Organisation eines Benutzers im Verhältnis zu anderen Auswirkungen anzupassen. ',
-    AR_X_Label:
-      'Die Zuweisung dieses Werts weist darauf hin, dass nicht genügend Informationen vorhanden sind, um einen der anderen Werte auszuwählen, und hat keinen Einfluss auf die Gesamtbewertung für die Umwelt, d. h. es hat die gleiche Auswirkung auf die Bewertung wie die Zuweisung von Mittel.',
-    AR_L_Label:
-      'Der Verlust der Verfügbarkeit hat wahrscheinlich nur begrenzte nachteilige Auswirkungen auf die Organisation oder Personen, die mit der Organisation verbunden sind (z. B. Mitarbeiter, Kunden).',
-    AR_M_Label:
-      'Das Zuweisen dieses Werts zur Metrik hat keinen Einfluss auf die Punktzahl.',
-    AR_H_Label:
-      'Der Verlust der Verfügbarkeit hat wahrscheinlich katastrophale nachteilige Auswirkungen auf die Organisation oder Personen, die mit der Organisation verbunden sind (z. B. Mitarbeiter, Kunden).',
-    MAV_Heading:
-      'Diese Metrik spiegelt den Kontext wider, in dem die Ausnutzung von Schwachstellen möglich ist. ',
-    MAV_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MAV_N_Label:
-      'Die verwundbare Komponente ist an den Netzwerkstack gebunden und die Menge möglicher Angreifer erstreckt sich über die anderen aufgeführten Optionen hinaus bis hin zum gesamten Internet. ',
-    MAV_A_Label:
-      'Die verwundbare Komponente ist an den Netzwerkstapel gebunden, aber der Angriff ist auf Protokollebene auf eine logisch benachbarte Topologie beschränkt. ',
-    MAV_L_Label:
-      'Die verwundbare Komponente ist nicht an den Netzwerkstapel gebunden und der Weg des Angreifers führt über Lese-/Schreib-/Ausführungsfunktionen. ',
-    MAV_P_Label:
-      'Der Angriff erfordert, dass der Angreifer die verwundbare Komponente physisch berührt oder manipuliert. ',
-    MAC_Heading:
-      'Diese Metrik beschreibt die Bedingungen, die außerhalb der Kontrolle des Angreifers liegen müssen, um die Schwachstelle auszunutzen. ',
-    MAC_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MAC_L_Label:
-      'Besondere Zugangsbedingungen oder mildernde Umstände bestehen nicht. ',
-    MAC_H_Label:
-      'Ein erfolgreicher Angriff hängt von Bedingungen ab, die außerhalb der Kontrolle des Angreifers liegen. ',
-    MPR_Heading:
-      'Diese Metrik beschreibt die Berechtigungsstufe, die ein Angreifer besitzen muss, bevor er die Schwachstelle erfolgreich ausnutzen kann.',
-    MPR_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MPR_N_Label:
-      'Der Angreifer ist vor dem Angriff nicht autorisiert und benötigt daher keinen Zugriff auf Einstellungen oder Dateien, um einen Angriff durchzuführen.',
-    MPR_L_Label:
-      'Der Angreifer ist mit (d. h. erfordert) Privilegien autorisiert, die grundlegende Benutzerfunktionen bereitstellen, die normalerweise nur Einstellungen und Dateien betreffen, die einem Benutzer gehören. ',
-    MPR_H_Label:
-      'Der Angreifer ist mit (d. h. erforderlichen) Privilegien autorisiert, die eine erhebliche (z. B. administrative) Kontrolle über die anfällige Komponente bieten, die sich auf komponentenweite Einstellungen und Dateien auswirken könnte.',
-    MUI_Heading:
-      'Diese Metrik erfasst die Anforderung an einen anderen Benutzer als den Angreifer, sich an der erfolgreichen Kompromittierung der anfälligen Komponente zu beteiligen. ',
-    MUI_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MUI_N_Label:
-      'Das anfällige System kann ohne Interaktion eines Benutzers ausgenutzt werden.',
-    MUI_R_Label:
-      'Die erfolgreiche Ausnutzung dieser Schwachstelle erfordert, dass ein Benutzer einige Maßnahmen ergreift, bevor die Schwachstelle ausgenutzt werden kann.',
-    MS_Heading:
-      'Wirkt sich ein erfolgreicher Angriff auf eine andere Komponente als die anfällige Komponente aus? ',
-    MS_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MS_U_Label:
-      'Eine ausgenutzte Schwachstelle kann nur Ressourcen betreffen, die von derselben Sicherheitsbehörde verwaltet werden. ',
-    MS_C_Label:
-      'Eine ausgenutzte Schwachstelle kann Ressourcen beeinträchtigen, die über den Sicherheitsbereich hinausgehen, der von der Sicherheitsautorität der anfälligen Komponente verwaltet wird. ',
-    MC_Heading:
-      'Diese Metrik misst die Auswirkungen auf die Vertraulichkeit der von einer Softwarekomponente verwalteten Informationsressourcen aufgrund einer erfolgreich ausgenutzten Schwachstelle. ',
-    MC_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MC_N_Label:
-      'Es gibt keinen Verlust der Vertraulichkeit innerhalb der betroffenen Komponente.',
-    MC_L_Label: 'Es gibt einen gewissen Verlust an Vertraulichkeit. ',
-    MC_H_Label:
-      'Es kommt zu einem vollständigen Verlust der Vertraulichkeit, was dazu führt, dass alle Ressourcen innerhalb der betroffenen Komponente an den Angreifer weitergegeben werden. ',
-    MI_Heading:
-      'Diese Metrik misst die Auswirkung auf die Integrität einer erfolgreich ausgenutzten Schwachstelle. ',
-    MI_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MI_N_Label:
-      'Es gibt keinen Integritätsverlust innerhalb der betroffenen Komponente.',
-    MI_L_Label:
-      'Die Änderung von Daten ist möglich, aber der Angreifer hat keine Kontrolle über die Folgen einer Änderung oder der Umfang der Änderung ist begrenzt. ',
-    MI_H_Label:
-      'Es kommt zu einem totalen Integritätsverlust oder einem kompletten Schutzverlust. ',
-    MA_Heading:
-      'Diese Metrik misst die Auswirkung auf die Verfügbarkeit der betroffenen Komponente, die sich aus einer erfolgreich ausgenutzten Schwachstelle ergibt. ',
-    MA_X_Label:
-      'Der der entsprechenden Basismetrik zugewiesene Wert wird verwendet.',
-    MA_N_Label:
-      'Es gibt keine Auswirkungen auf die Verfügbarkeit innerhalb der betroffenen Komponente.',
-    MA_L_Label:
-      'Die Leistung wird reduziert oder es gibt Unterbrechungen in der Ressourcenverfügbarkeit. ',
-    MA_H_Label:
-      'Es kommt zu einem vollständigen Verlust der Verfügbarkeit, was dazu führt, dass der Angreifer den Zugriff auf Ressourcen in der betroffenen Komponente vollständig verweigern kann; '
-  },
   report_type: {
     add_new: 'Neuen Berichtselementtyp hinzufügen',
     edit: 'Berichtselementtyp bearbeiten',
